+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/graph/model/AnyItem.java
@@ -1,0 +1,28 @@
+package com.jpexs.decompiler.graph.model;
+
+import com.jpexs.decompiler.flash.helpers.GraphTextWriter;
+import com.jpexs.decompiler.graph.GraphTargetItem;
+import com.jpexs.decompiler.graph.TypeItem;
+
+/**
+ *
+ * @author JPEXS
+ */
+public class AnyItem extends GraphTargetItem {
+
+    @Override
+    public GraphTextWriter appendTo(GraphTextWriter writer, LocalData localData) throws InterruptedException {
+        return writer.append("§§any()");
+    }
+
+    @Override
+    public boolean hasReturnValue() {
+        return true;
+    }
+
+    @Override
+    public GraphTargetItem returnType() {
+        return TypeItem.UNBOUNDED;
+    }
+
+}