+++ CHANGELOG.md
@@ -24,6 +24,7 @@ All notable changes to this project will be documented in this file.
 - Improper initialization of ActiveX component when Flash not available causing FFDec not start
 - #1206 Switch with multiple default clauses
 - ASC2 §§push of function calls before returning from a method
+- Support for ASC2 and swftools try..catch..finally block
 
 ### Changed
 - AS3 test methods separated to classes